+++ macros/Plotting/PlotUtilities.h
@@ -29,10 +29,10 @@ TH1F* getHist1D(string title, std::map<string, TH1F*> &allhistos,
 void plot2D(string title, float xval, float yval, double weight, std::map<string, TH2F*> &allhistos, 
 	    int numbinsx, float xmin, float xmax, int numbinsy, float ymin, float ymax);
 
-void plot1DUnderOverFlow(string title, float xval, double weight, std::map<string, TH1D*> &allhistos, 
-	    int numbinsx, float xmin, float xmax);
-void plot2DUnderOverFlow(string title, float xval, float yval, double weight, std::map<string, TH2D*> &allhistos, 
-	    int numbinsx, float xmin, float xmax, int numbinsy, float ymin, float ymax);
+void plot1DUnderOverFlow(string title, double xval, double weight, std::map<string, TH1D*> &allhistos, 
+	    int numbinsx, double xmin, double xmax);
+void plot2DUnderOverFlow(string title, double xval, double yval, double weight, std::map<string, TH2D*> &allhistos, 
+	    int numbinsx, double xmin, double xmax, int numbinsy, double ymin, double ymax);
 
 void savePlots(std::map<string, TH1F*>&, char* );
 void savePlots2(std::map<string, TH2F*>&, char* );
